+++ bench/addictive_calling/addictive_calling.java
@@ -1,10 +1,10 @@
 class AddictiveCalling {
-    static int increment(int x) {
+    static i64 increment(i64 x) {
         return x + 1;
     }
 
     public static void main(String[] args) {
-        int i = 0;
+        i64 i = 0;
         while (i < 10_000_000) {
             i = increment(i);
         }
